added source and windows exe.
for additional details, BRUHIFF only works on windows, so this only works on windows. (unless you want to download python and run the .py file although it is preinstalled on most linux and mac machines, you should still probably get the latest version of python from the website, so really it is useless. unless you want to see the source code.)